+++ src/main/java/odk/groupe4/ApiCollabDev/controllers/ContributeurController.java
@@ -5,6 +5,7 @@ import odk.groupe4.ApiCollabDev.dto.ProjetDto;
 import odk.groupe4.ApiCollabDev.models.Contributeur;
 import odk.groupe4.ApiCollabDev.models.enums.StatusProject;
 import odk.groupe4.ApiCollabDev.service.ContributeurSercice;
+import odk.groupe4.ApiCollabDev.service.ProjetService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,6 +13,7 @@ import org.springframework.web.bind.annotation.*;
 
 public class ContributeurController {
     @Autowired
+    private ProjetService projetService;
     private ContributeurSercice contributeurSercice;
 
     @PostMapping
@@ -22,11 +24,11 @@ public class ContributeurController {
     //Méthode pour proposer un projet
     @PostMapping("/proposerProjets")
     public ProjetDto proposerProjet(@RequestBody ProjetDto projetDto) {
-        return contributeurSercice.proposerProjet(projetDto);
+        return projetService.proposerProjet(projetDto);
     }
     // Méthode pour recuperer l'avancement d'un projet
     @GetMapping("/projets/{id}/avancement")
     public StatusProject recupererAvancementProjet(@PathVariable int id) {
-        return contributeurSercice.suivreAvancementProjet(id);
+        return projetService.suivreAvancementProjet(id);
     }
 }
