+++ src/main/java/odk/groupe4/ApiCollabDev/models/enums/Profil.java
@@ -1,6 +1,7 @@
 package odk.groupe4.ApiCollabDev.models.enums;
 
 public enum Profil {
+    PORTEUR_DE_PROJET,
     DEVELOPPEUR,
     DESIGNER,
     GESTIONNAIRE
